changed tests creation and view so the logged in teacher will only be able to create tests for his courses and view tests for his courses. didnt run it yet.

+++ TestUs/src/main/java/com/springbootquickstart/TestUs/repository/StudentRepository.java
@@ -11,8 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
     Optional<Student> findByMyUser(MyUser myUser);
-
 
 }
